# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DocArchiv",
    "description": "Leichtgewichtiger Dokumenten-Katalog mit Volltextsuche und Nextcloud-Linkout.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "description": "Liefert Dokumente mit Suche, Filtern, Pagination und Sortierung.",
        "operationId": "list_documents_api_documents_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Volltextsuche",
              "title": "Q"
            },
            "description": "Volltextsuche"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kommaseparierte Tag-Liste",
              "title": "Tags"
            },
            "description": "Kommaseparierte Tag-Liste"
          },
          {
            "name": "document_type_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Dokumenttyp-ID",
              "title": "Document Type Id"
            },
            "description": "Dokumenttyp-ID"
          },
          {
            "name": "correspondent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Korrespondent-ID",
              "title": "Correspondent Id"
            },
            "description": "Korrespondent-ID"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Startdatum inklusive",
              "title": "Date From"
            },
            "description": "Startdatum inklusive"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Enddatum inklusive",
              "title": "Date To"
            },
            "description": "Enddatum inklusive"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Per Page"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "date_desc",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_DocumentListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Create Document",
        "description": "Legt ein neues Dokument an.",
        "operationId": "create_document_api_documents_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "description": "Liefert ein einzelnes Dokument anhand seiner ID.",
        "operationId": "get_document_api_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "documents"
        ],
        "summary": "Update Document",
        "description": "Aktualisiert ein vorhandenes Dokument teilweise.",
        "operationId": "update_document_api_documents__document_id__put",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Delete Document",
        "description": "Loescht ein Dokument.",
        "operationId": "delete_document_api_documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/document-types": {
      "get": {
        "tags": [
          "document-types"
        ],
        "summary": "List Document Types",
        "description": "Liefert alle Dokumenttypen inklusive Dokumentanzahl.",
        "operationId": "list_document_types_api_document_types_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentTypeResponse"
                  },
                  "type": "array",
                  "title": "Response List Document Types Api Document Types Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "document-types"
        ],
        "summary": "Create Document Type",
        "description": "Legt einen neuen Dokumenttyp an.",
        "operationId": "create_document_type_api_document_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentTypeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentTypeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/document-types/{document_type_id}": {
      "get": {
        "tags": [
          "document-types"
        ],
        "summary": "Get Document Type",
        "description": "Liefert einen einzelnen Dokumenttyp anhand seiner ID.",
        "operationId": "get_document_type_api_document_types__document_type_id__get",
        "parameters": [
          {
            "name": "document_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentTypeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "document-types"
        ],
        "summary": "Update Document Type",
        "description": "Aktualisiert einen vorhandenen Dokumenttyp teilweise.",
        "operationId": "update_document_type_api_document_types__document_type_id__patch",
        "parameters": [
          {
            "name": "document_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Type Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentTypeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentTypeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "document-types"
        ],
        "summary": "Delete Document Type",
        "description": "Loescht einen Dokumenttyp.",
        "operationId": "delete_document_type_api_document_types__document_type_id__delete",
        "parameters": [
          {
            "name": "document_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Type Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "original_filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Original Filename"
          },
          "stored_filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Stored Filename"
          },
          "document_type_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Document Type Id"
          },
          "correspondent_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Correspondent Id"
          },
          "document_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Date"
          },
          "nextcloud_path": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Nextcloud Path"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "title",
          "original_filename",
          "stored_filename",
          "document_type_id",
          "nextcloud_path"
        ],
        "title": "DocumentCreate",
        "description": "Request-Schema zum Anlegen eines Dokuments durch die Scan-Pipeline."
      },
      "DocumentListResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "stored_filename": {
            "type": "string",
            "title": "Stored Filename"
          },
          "document_type_id": {
            "type": "integer",
            "title": "Document Type Id"
          },
          "correspondent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correspondent Id"
          },
          "document_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Date"
          },
          "nextcloud_path": {
            "type": "string",
            "title": "Nextcloud Path"
          },
          "nextcloud_url": {
            "type": "string",
            "title": "Nextcloud Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "document_type": {
            "$ref": "#/components/schemas/DocumentTypeResponse"
          },
          "correspondent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CorrespondentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagResponse"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "summary",
          "original_filename",
          "stored_filename",
          "document_type_id",
          "correspondent_id",
          "document_date",
          "nextcloud_path",
          "nextcloud_url",
          "created_at",
          "updated_at",
          "document_type"
        ],
        "title": "DocumentListResponse",
        "description": "Reduziertes Response-Schema fuer Dokumentlisten."
      },
      "DocumentResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "original_filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Original Filename"
          },
          "stored_filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Stored Filename"
          },
          "document_type_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Document Type Id"
          },
          "correspondent_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Correspondent Id"
          },
          "document_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Date"
          },
          "nextcloud_path": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Nextcloud Path"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nextcloud_url": {
            "type": "string",
            "title": "Nextcloud Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "document_type": {
            "$ref": "#/components/schemas/DocumentTypeResponse"
          },
          "correspondent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CorrespondentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagResponse"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "title",
          "original_filename",
          "stored_filename",
          "document_type_id",
          "nextcloud_path",
          "id",
          "nextcloud_url",
          "created_at",
          "updated_at",
          "document_type"
        ],
        "title": "DocumentResponse",
        "description": "Response-Schema fuer ein Dokument."
      },
      "DocumentTypeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DocumentTypeCreate",
        "description": "Request-Schema zum Anlegen eines Dokumenttyps."
      },
      "DocumentTypeResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "DocumentTypeResponse",
        "description": "Response-Schema fuer Dokumenttypen."
      },
      "DocumentTypeUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          }
        },
        "type": "object",
        "title": "DocumentTypeUpdate",
        "description": "Request-Schema fuer teilweise Dokumenttyp-Updates."
      },
      "DocumentUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "original_filename": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Filename"
          },
          "stored_filename": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Stored Filename"
          },
          "document_type_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Type Id"
          },
          "correspondent_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Correspondent Id"
          },
          "document_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Date"
          },
          "nextcloud_path": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcloud Path"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "DocumentUpdate",
        "description": "Request-Schema fuer teilweise Dokument-Updates aus der SPA."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_DocumentListResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentListResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Per Page"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page",
          "pages"
        ],
        "title": "PaginatedResponse[DocumentListResponse]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}